+++ example/android/app/src/main/kotlin/com/example/template_fluttter/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.template_fluttter
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()