+++ example/integration_test/app_features.dart
@@ -3,10 +3,12 @@
 
 // clear && printf '\e[3J' && flutter drive -t integration_test/app_features.dart
 
+// clear && printf '\e[3J' && flutter run -t integration_test/app_features.dart
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test_preview/integration_test_binding.dart';
-import 'package:device_frame_community/src/devices/devices.dart';
-import 'package:device_frame_community/src/info/info.dart';
+import 'package:device_frame/src/devices/devices.dart';
+import 'package:device_frame/src/info/info.dart';
 
 import 'package:example/main.dart' as app;
 import 'app_feature_groups.dart';
@@ -15,7 +17,7 @@ void main() async {
 
     const minutesPerDevice = 7;
     final Set<DeviceInfo> testDevices = {
-        Devices.ios.iPadPro11Inches, Devices.android.samsungGalaxyNote20, Devices.ios.iPhone12
+        Devices.ios.iPadPro11Inches, Devices.android.samsungGalaxyNote20
     };
     final totalExpectedDuration = Duration(minutes: testDevices.length * minutesPerDevice);
     final binding = IntegrationTestPreviewBinding.ensureInitialized();