+++ test/integration_test_preview_test.dart
@@ -1,4 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:integration_test_preview/integration_test_preview.dart';
-